♻️ Refactor: 모바일 폴더 구조 정리
