+++ apps/mobile/src/screens/home/HomeScreen.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react-native';
 
-import HomeScreen from '../src/app/index';
+import HomeScreen from './HomeScreen';
 
 describe('<HomeScreen />', () => {
   it('웹 주소가 설정되지 않으면 안내 문구를 보여준다', async () => {
